+++ src/main/java/com/ssrprojects/ultimatechatapp/controller/AuthenticationController.java
@@ -4,9 +4,9 @@ package com.ssrprojects.ultimatechatapp.controller;
 import com.ssrprojects.ultimatechatapp.config.security.JWTUtil;
 import com.ssrprojects.ultimatechatapp.service.UserService.UserService;
 import jakarta.validation.Valid;
-import model.LoginRequest;
-import model.SignUpRequest;
-import model.VerificationRequest;
+import com.ssrprojects.ultimatechatapp.model.LoginRequest;
+import com.ssrprojects.ultimatechatapp.model.SignUpRequest;
+import com.ssrprojects.ultimatechatapp.model.VerificationRequest;
 import org.springframework.data.util.Pair;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
