+++ src/main/java/com/ssrprojects/ultimatechatapp/model/VerificationRequest.java
@@ -1,4 +1,4 @@
-package model;
+package com.ssrprojects.ultimatechatapp.model;
 
 import lombok.Getter;
